// typechain-types/NFTMarketplace.ts
export declare namespace NFTMarketplace {
  export type MarketItemStruct = {
    tokenId: BigNumberish;
    seller: string;
    owner: string;
    price: BigNumberish;
    sold: boolean;
  };

  export type MarketItemStructOutput = [
    BigNumber,
    string,
    string,
    BigNumber,
    boolean
  ] & {
    tokenId: BigNumber;
    seller: string;
    owner: string;
    price: BigNumber;
    sold: boolean;
  };
}